+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Current (in progress)
 
 - Only store `url` field in `remote_url` extra if this is an URL otherwise store it in `ckan:source` [#30](https://github.com/opendatateam/udata-ckan/pull/30)
+- Properly handle geometry errors [#31](https://github.com/opendatateam/udata-ckan/pull/31)
 
 ## 1.1.0 (2018-06-06)