+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Current (in progress)
+## 1.1.1 (2018-06-15)
 
 - Only store `url` field in `remote_url` extra if this is an URL otherwise store it in `ckan:source` [#30](https://github.com/opendatateam/udata-ckan/pull/30)
 - Properly handle geometry errors [#31](https://github.com/opendatateam/udata-ckan/pull/31)